+++ src/main/java/com/yunhuakeji/attendance/enums/CareStatus.java
@@ -4,18 +4,18 @@ package com.yunhuakeji.attendance.enums;
  * 关怀状态枚举
  */
 public enum CareStatus {
-    NO(1, "未关怀"),
-    YES(2, "已关怀");
+    NO((byte)1, "未关怀"),
+    YES((byte)2, "已关怀");
 
-    private int type;
+    private byte type;
     private String msg;
 
-    CareStatus(int type, String msg) {
+    CareStatus(byte type, String msg) {
         this.type = type;
         this.msg = msg;
     }
 
-    public static CareStatus get(int val) {
+    public static CareStatus get(byte val) {
         for (CareStatus c : CareStatus.values()) {
             if (c.getType() == val) {
                 return c;
@@ -24,7 +24,7 @@ public enum CareStatus {
         return null;
     }
 
-    public int getType() {
+    public byte getType() {
         return type;
     }
